Wrap owning File* in unique_ptrs
The C++ Core Guidelines suggest to use std::unique_ptr<T> instead of T*
to express ownership of T.
This way the objects get deallocated automatically when the unique_ptrs
are destroyed.
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rf-unique_ptr

The class Setup contains a std::map mFileMap that owns the files and additionally
two observing vectors inputFiles/outputFiles, which only refer to the std::map entries.